refactor: keep "Not"-expectations in same file
