+++ Source/aweXpect/That/Collections/ThatDictionaryShould.ContainKey.cs
@@ -19,6 +19,18 @@ public static partial class ThatDictionaryShould
 			source
 		);
 
+	/// <summary>
+	///     Verifies that the dictionary does not contain the <paramref name="unexpected" /> key.
+	/// </summary>
+	public static AndOrResult<IDictionary<TKey, TValue>, IThat<IDictionary<TKey, TValue>>> NotContainKey<TKey, TValue>(
+		this IThat<IDictionary<TKey, TValue>> source,
+		TKey unexpected)
+		=> new(
+			source.ExpectationBuilder.AddConstraint(it
+				=> new NotContainKeyConstraint<TKey, TValue>(it, unexpected)),
+			source
+		);
+
 	private readonly struct ContainKeyConstraint<TKey, TValue>(string it, TKey expected)
 		: IValueConstraint<IDictionary<TKey, TValue>>
 	{
@@ -41,5 +53,29 @@ public static partial class ThatDictionaryShould
 		}
 
 		public override string ToString() => $"have key {Formatter.Format(expected)}";
+	}
+
+	private readonly struct NotContainKeyConstraint<TKey, TValue>(string it, TKey unexpected)
+		: IValueConstraint<IDictionary<TKey, TValue>>
+	{
+		public ConstraintResult IsMetBy(IDictionary<TKey, TValue> actual)
+		{
+			// ReSharper disable once ConditionIsAlwaysTrueOrFalseAccordingToNullableAPIContract
+			if (actual is null)
+			{
+				return new ConstraintResult.Failure(ToString(),
+					$"{it} was <null>");
+			}
+
+			if (actual.ContainsKey(unexpected))
+			{
+				return new ConstraintResult.Failure<IDictionary<TKey, TValue>>(actual, ToString(),
+					$"{it} did");
+			}
+
+			return new ConstraintResult.Success<IDictionary<TKey, TValue>>(actual, ToString());
+		}
+
+		public override string ToString() => $"not have key {Formatter.Format(unexpected)}";
 	}
 }
