+++ Source/aweXpect/That/Collections/ThatDictionaryShould.ContainKeys.cs
@@ -20,6 +20,18 @@ public static partial class ThatDictionaryShould
 			source
 		);
 
+	/// <summary>
+	///     Verifies that the dictionary contains none of the <paramref name="unexpected" /> keys.
+	/// </summary>
+	public static AndOrResult<IDictionary<TKey, TValue>, IThat<IDictionary<TKey, TValue>>> NotContainKeys<TKey, TValue>(
+		this IThat<IDictionary<TKey, TValue>> source,
+		params TKey[] unexpected)
+		=> new(
+			source.ExpectationBuilder.AddConstraint(it
+				=> new NotContainKeysConstraint<TKey, TValue>(it, unexpected)),
+			source
+		);
+
 	private readonly struct ContainKeysConstraint<TKey, TValue>(string it, TKey[] expected)
 		: IValueConstraint<IDictionary<TKey, TValue>>
 	{
@@ -43,5 +55,30 @@ public static partial class ThatDictionaryShould
 		}
 
 		public override string ToString() => $"have keys {Formatter.Format(expected)}";
+	}
+
+	private readonly struct NotContainKeysConstraint<TKey, TValue>(string it, TKey[] unexpected)
+		: IValueConstraint<IDictionary<TKey, TValue>>
+	{
+		public ConstraintResult IsMetBy(IDictionary<TKey, TValue> actual)
+		{
+			// ReSharper disable once ConditionIsAlwaysTrueOrFalseAccordingToNullableAPIContract
+			if (actual is null)
+			{
+				return new ConstraintResult.Failure(ToString(),
+					$"{it} was <null>");
+			}
+
+			List<TKey> existingKeys = unexpected.Where(actual.ContainsKey).ToList();
+			if (existingKeys.Any())
+			{
+				return new ConstraintResult.Failure<IDictionary<TKey, TValue>>(actual, ToString(),
+					$"{it} did have {Formatter.Format(existingKeys, FormattingOptions.MultipleLines)}");
+			}
+
+			return new ConstraintResult.Success<IDictionary<TKey, TValue>>(actual, ToString());
+		}
+
+		public override string ToString() => $"not have keys {Formatter.Format(unexpected)}";
 	}
 }
