+++ Source/aweXpect/That/Collections/ThatDictionaryShould.ContainValue.cs
@@ -19,6 +19,19 @@ public static partial class ThatDictionaryShould
 			source
 		);
 
+	/// <summary>
+	///     Verifies that the dictionary does not contain the <paramref name="unexpected" /> value.
+	/// </summary>
+	public static AndOrResult<IDictionary<TKey, TValue>, IThat<IDictionary<TKey, TValue>>>
+		NotContainValue<TKey, TValue>(
+			this IThat<IDictionary<TKey, TValue>> source,
+			TValue unexpected)
+		=> new(
+			source.ExpectationBuilder.AddConstraint(it
+				=> new NotContainValueConstraint<TKey, TValue>(it, unexpected)),
+			source
+		);
+
 	private readonly struct ContainValueConstraint<TKey, TValue>(string it, TValue expected)
 		: IValueConstraint<IDictionary<TKey, TValue>>
 	{
@@ -41,5 +54,29 @@ public static partial class ThatDictionaryShould
 		}
 
 		public override string ToString() => $"have value {Formatter.Format(expected)}";
+	}
+
+	private readonly struct NotContainValueConstraint<TKey, TValue>(string it, TValue unexpected)
+		: IValueConstraint<IDictionary<TKey, TValue>>
+	{
+		public ConstraintResult IsMetBy(IDictionary<TKey, TValue> actual)
+		{
+			// ReSharper disable once ConditionIsAlwaysTrueOrFalseAccordingToNullableAPIContract
+			if (actual is null)
+			{
+				return new ConstraintResult.Failure(ToString(),
+					$"{it} was <null>");
+			}
+
+			if (actual.ContainsValue(unexpected))
+			{
+				return new ConstraintResult.Failure<IDictionary<TKey, TValue>>(actual, ToString(),
+					$"{it} did");
+			}
+
+			return new ConstraintResult.Success<IDictionary<TKey, TValue>>(actual, ToString());
+		}
+
+		public override string ToString() => $"not have value {Formatter.Format(unexpected)}";
 	}
 }
