--- Source/aweXpect/That/Collections/ThatDictionaryShould.NotContainValue.cs
@@ -1,45 +1,0 @@
-﻿using System.Collections.Generic;
-using aweXpect.Core;
-using aweXpect.Core.Constraints;
-using aweXpect.Results;
-
-namespace aweXpect;
-
-public static partial class ThatDictionaryShould
-{
-	/// <summary>
-	///     Verifies that the dictionary does not contain the <paramref name="unexpected" /> value.
-	/// </summary>
-	public static AndOrResult<IDictionary<TKey, TValue>, IThat<IDictionary<TKey, TValue>>> NotContainValue<TKey, TValue>(
-		this IThat<IDictionary<TKey, TValue>> source,
-		TValue unexpected)
-		=> new(
-			source.ExpectationBuilder.AddConstraint(it
-				=> new NotContainValueConstraint<TKey, TValue>(it, unexpected)),
-			source
-		);
-
-	private readonly struct NotContainValueConstraint<TKey, TValue>(string it, TValue unexpected)
-		: IValueConstraint<IDictionary<TKey, TValue>>
-	{
-		public ConstraintResult IsMetBy(IDictionary<TKey, TValue> actual)
-		{
-			// ReSharper disable once ConditionIsAlwaysTrueOrFalseAccordingToNullableAPIContract
-			if (actual is null)
-			{
-				return new ConstraintResult.Failure(ToString(),
-					$"{it} was <null>");
-			}
-
-			if (actual.ContainsValue(unexpected))
-			{
-				return new ConstraintResult.Failure<IDictionary<TKey, TValue>>(actual, ToString(),
-					$"{it} did");
-			}
-
-			return new ConstraintResult.Success<IDictionary<TKey, TValue>>(actual, ToString());
-		}
-
-		public override string ToString() => $"not have value {Formatter.Format(unexpected)}";
-	}
-}
